move metrics to datagen temporarily
