+++ inference_perf/datagen/base.py
@@ -13,8 +13,7 @@
 # limitations under the License.
 import aiohttp
 import numpy as np
-from inference_perf.metrics.base import Metric, MetricsCollector
-from inference_perf.reportgen.base import ReportFile
+from inference_perf.reportgen.base import Metric, MetricsCollector, ReportFile
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
 from pydantic import BaseModel
 from inference_perf.config import APIType, Distribution, RequestLifecycleMetricsReportConfig
