+++ inference_perf/metrics/base.py
@@ -12,35 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from abc import ABC, abstractmethod
-from typing import Any, Generic, List, Optional, TypeVar
-
-from pydantic import BaseModel
 from inference_perf.client.modelserver.base import ModelServerClient, ModelServerMetrics
-from inference_perf.reportgen.base import ReportFile
-
-
-class Metric(BaseModel):
-    """Abstract type to track reportable (but not neccesarily summarizable) metrics"""
-
-    stage_id: Optional[int] = None
-
-    @abstractmethod
-    async def to_report(self) -> dict[str, Any]:
-        """Create the report for this metric"""
-        raise NotImplementedError
-
-
-T = TypeVar("T", bound=Metric)
-
-
-class MetricsCollector(ABC, BaseModel, Generic[T]):
-    """Anything that can collect metrics to be included in the output report"""
-
-    metrics: List[T]
-
-    @abstractmethod
-    async def to_report(self, report_config: Any) -> List[ReportFile]:
-        raise NotImplementedError
 
 
 class PerfRuntimeParameters:
