+++ inference_perf/reportgen/base.py
@@ -11,12 +11,39 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from abc import ABC, abstractmethod
 import json
 import statistics
-from typing import Any, List
+from typing import Any, Generic, List, Optional, TypeVar
+
+from pydantic import BaseModel
 from inference_perf.metrics import MetricsClient
 from inference_perf.client.modelserver.base import ModelServerMetrics
 from inference_perf.metrics.base import PerfRuntimeParameters
+
+
+class Metric(BaseModel):
+    """Abstract type to track reportable (but not neccesarily summarizable) metrics"""
+
+    stage_id: Optional[int] = None
+
+    @abstractmethod
+    async def to_report(self) -> dict[str, Any]:
+        """Create the report for this metric"""
+        raise NotImplementedError
+
+
+T = TypeVar("T", bound=Metric)
+
+
+class MetricsCollector(ABC, BaseModel, Generic[T]):
+    """Anything that can collect metrics to be included in the output report"""
+
+    metrics: List[T]
+
+    @abstractmethod
+    async def to_report(self, report_config: Any) -> List["ReportFile"]:
+        raise NotImplementedError
 
 
 class ReportFile:
